requirements.txt
functions-framework==3.*
requests==2.28.*
google-auth==1.35.*
google-cloud==0.34.*
google-cloud-bigquery==1.28.*
pytz==2022.7.*
numpy==1.19.*
pandas==1.1.5
